requirements.txt
absl-py
ftfy
gdown==3.13.1
gym
hydra-core==1.0.5
imageio==2.9.0
imageio-ffmpeg==0.4.5
kornia==0.4.1
matplotlib
meshcat==0.0.18
numpy
opencv-python
pybullet
pytorch_lightning
regex
scikit-image==0.17.2
scipy
tdqm
torch
torchvision
transformers==4.3.2
transforms3d==0.3.1
wandb
airobot
